requirements.txt
altgraph==0.17.4
arabic-reshaper==3.0.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.58.4
fpdf2==2.8.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
gunicorn


